Change signal field name to 'sample_frequency'
The terms 'sample_rate' and 'sample_frequency' were used
interchangeably in the codebase, which is valid as long as the
record duration is 1 second, but they stop being the same when
the record duration is any other value. I followed the meaning
established in the documentation, where 'sample_frequency' is
taken to represent the number of samples per signal per record,
and 'sample_rate' is the number samples per signal per second.

This represents a breaking change to pyedflib's API, as it changes
the name of various EdfWriter method parameters and keys
in the signal dictionaries returned by EdfReader methods.
However, it conforms to the terminology used in the EDF spec
(that the header displays the number of samples in each data
record) as well as edflib function names. It also partially solves
the mismatched header values when record duration isn't 1 second.