Make addition of 'sample_frequency' backwards compatible
This is to avoid a breaking change with the addition of the
'sample_frequency' param in order to allow a deprecation notice
period. This will permit the current change to be part of a
patch release, with the removal of the `sample_rate` param likely
being saved for a major release.